pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.3"
python-dotenv = "^0.14.0"
requests = "^2.28.2"
pytest = "^7.2.2"
selenium = "^4.8.2"
gunicorn = "^20.1.0"
pymongo = "^4.5.0"
mongomock = "^4.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
